[ios-setup] Copy the project settings into the the .xcconfig files.
This copies the Xcode project settings into the relative .xcconfig files.
As no changes have been made to this project, these are the default project
values. To copy these settings into the .xcconfig files, select the Project
settings in Xcode and go the Build Settings. Be sure that you selected Project
and not Targets. Then be sure to select All and Levels. This will reveal
all values that differ from their default values and will show you the
hierarchy in which the settings are applied. The default values
are shown in the right-most column titled iOS Default. Select Project and note
the hierarchy, it will be:
Resolved -> Project Name -> Config.File -> iOS Default

Next highlight a value and copy it with CMD+C, and then paste it with CMD+V
into the respective .xcconfig file.
- Settings that are equal in all Debug, Test, Stage, and Release are set in Global.xcconfig.
- Settings unique to the Debug configuration are set in Debug.xcconfig.
- Settings unique to the Test configuration are set in Test.xcconfig.
- Settings unique to the Stage configuration are set in Stage.xcconfig.
- Settings unique to the Release configuration are set in Release.xcconfig.
Note: At this point, Debug.xcconfig and Test.xcconfig should be the same,
and Stage.xcconfig and Release.xcconfig should be the same.

Now that the settings are in configuration files, it is easier to maintain
them via configuration management. It also allows for an easy way to test
a new setting; by updating the project settings directly in Xcode, a new
setting value can be tested without losing the setting stored in the
.xcconfig file. Permanent settings changes would then be moved to the
correct settings file.