[ios-setup] Copy the target settings into the the .xcconfig files.
This copies the Xcode target settings into the relative .xcconfig files.
As no changes have been made to this project, these are the default target
values. To copy these settings into the .xcconfig files, select the Targets
settings in Xcode and go the Build Settings. Be sure that you selected Targets
and not Project. Then be sure to select All and Levels. This will reveal
all values that differ from their default values and will show you the
hierarchy in which the settings are applied. The default values
are shown in the right-most column titled iOS Default. The hierarchy is:
Resolved -> project name (target) -> Config.File (target) -> project name (project) -> Config.File (project) -> iOS Default

Next highlight a value and copy it with CMD+C, and then paste it with CMD+V
into the respective .xcconfig file.
- Copy Debug settings into Debug.xcconfig
- Copy Test settings into Test.xcconfig
- Copy Stage settings into Stage.xcconfig
- Copy Release settings into Release.xcconfig
Note: At this point, Debug.xcconfig and Test.xcconfig should be the same,
and Stage.xcconfig and Release.xcconfig should be the same.

Now the Target settings have been copied into their respective .xcconfig files.