[ios-setup] Change the 'product name' in the xcconfig files to match the environment.
Change the product name in the xcconfig file so that the:
- Debug build will prepend 'dev' to the product name
- Test build will prepend 'test' to the product name
- Stage build will prepend 'stage' to the product name

We leave the Release build alone, as adding 'prod' to the product name
is not very appealing.

Now changing the scheme will change the name of the product, where the name
is now controlled by the xcconfig file. This shows the xcconfig files working.