[ios-setup] Change the product name in the xcconfig files.
Move the environment name from the beginning of the product name
to the end of the product name. Now that the environment can be
identified via the app icon, it's nicer to have the name at the
end; in fact the name isn't really necessary as it is the
bundle identifier that creates the separate builds.