[plist-example] Add a Project Settings dictionary to Info.plist.
This adds a new dictionary entry called 'Project Settings' to Info.plist.
By creating a dictionary, we can store all our own values under this
dictionary for better organization and namespacing purposes. If we
just add new values, it becomes difficult to tell which values were
added by us, and which were created by Apple or a 3rd party.

We also add an 'Environment' entry where the value is $(Environment).
This tells Info.plist to get this value from the app's settings, and
of course, those are now configured in the xcconfig files.