[cocoapods] Add AFNetworking to the codebase.
This commit is simply to illustrate how to import AFNetworking, installed
through CocoaPods, into the project. Note that for this work, the
xcworkspace must be used intead of the xcodeproj file. The xcodeproj file
will state that it cannot find AFNetworking. A workspace is effectively a
series of projects, and now there is the original project as well as the
import CocoaPod project. Hence the workspace is now the 'project' file.