[cocoapods] Add the CocoaPods changes to the project.
This adds the changes that CocoaPods made to the project to the actual project file.
Now the newly created xcworkspace file can be considered a set of project files, and
of course the app project file needs to be included in that. As such, that project
file, which is this file, needs to have all the CocoaPods changes for the it to
know all the dependencies and work correctly.

In short, not committing these changes would cause problems in the project.