[swift-bridging] Create a Swift file and its bridging header file.
Create a Swift file named StartupSwiftExample.swift through Xcode and add it
to the Utils group. When Xcode asks you to add an Objective-C bridging Header,
choose 'Yes' and this will create the ios-startup-guide-Bridging-Header.h file.

The name of the bridging header is auto-generated based off the Product Name,
which is set to the Target Name. It appears in the Utils group simple because
the Swift file was created in the Utils group.