[swift-bridging] Remove the environment name from the Product name setting.
This updates the Product Name setting in Debug, Test, and Stage xcconfig files.
The reason for this is that to use Swift in an Objective-C file, you must
import ios_startup_guide-Swift.h, ie the Product Name and -Swift.h. With the
previous xcconfig files, there would have been four different import files, ie
ios_startup_guide_Dev-Swift.h, ios_startup_guide_Test-Swift.h, etc, and
to import the file would have required a series of IFDEFs. It's much easier to
set the Product Name to not use the environment in it. Using the Product Name
with the environment modifier was just aesthetics anyways, as the Bundle
identifier is what creates the separate builds. Besides, the icons are different
per environment and that serves as the means to distinguish between the
environments.