[remove-storyboards] Create three label properties on the StartupHomeViewController.m.
This creates three labels for the StartupHomeViewController through the
.m file's interface declaration. Generally view properties, such as
labels, text fields, etc, would not be modified by another controller,
so they are added to the .m file. This makes them private and inaccessible
outside of the StartupHomeViewController. Obviously if it was needed for
them to be public, they would be declared in the .h file, but it's better
to start by putting them in the .m file.