[remove-storyboards] Remove the 'Main storyboard file base name' entry from Info.plist.
The goal of this branch was to remove the storyboard to show how to use xib files instead.
This removes the UIMainStoryboardFile entry form the Info.plist. Now that the
AppDelegate loads the StartupHomeViewController, it is not needed. Storyboard files are
certainly easier to work with, but can be difficult to use in a team setting, particularly
as they are difficult to diff/merge should someone else modify the storyboard.