[remove-storyboards] Set the 'Launch screen interface file basename' in Info.plist.
This changes the 'Launch screen interface file basename' from being LaunchScreen.storyboard
to being StartupHomeView.xib. If the 'Launch screen interface file basename' is removed
from the Info.plist entirely, a warning appears stating that:
'A launch image named 'Default-568h@2x.png' is required to run at native resolution on
Retina 4-inch devices'.

To resolve that you can either add a Launch Icon to the Assets, or set it to use a
Storyboard or xib file. In this example we set it to use the StartupHomeView.xib file.
In all practicality, a default LaunchScreen.storyboard file will probably not cause a lot
merge conflicts as it would not be modified often, as opposed to the Main.storyboard.