[remove-storyboards] Reset the 'Launch screen interface file base name' in the Info.plist.
This resets the 'Launch screen interface file base name' back to being the
LaunchScreen.storyboard file. As we demonstrated how to remove this file
successfully, we also mentioned why it is easier to keep it. So now set it
back, but keep the commit history to show to not use it correctly.