[child-controller] Add and remove a child view in StartupHomeViewController.m.
This shows how to add a child view controller and remove a child view
controller in the StartupHomeViewController.m. Note that this example
is trivial, and this would be better done through presentViewController.
The point, though, is to show how it is done in case this technique is
needed. As mentioned here
https://stackoverflow.com/questions/9136391/addchildviewcontroller-and-presentviewcontroller
'The parent/child relationship is for when a view controller has subviews that are managed
by their own view controllers.'