[ios-setup] Create a 'Docs' directory for documentation support.
Create a 'Docs' directory along with an 'Images' subdirectory with the images needed
to support the README section that will outline how to create the xcconfig files.

This directory would also work as a location to store other markdown files if parts
of the README would be better suited in a separate file, perhaps for clarity or for
greater detail.